magic circle improvements

+++ Content/Projectiles/Enemies/DuskHound.cs
@@ -6,6 +6,7 @@ using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using Terraria;
 using Terraria.Audio;
+using Terraria.ID;
 using Terraria.ModLoader;
 
 namespace ArcaneOdyssey.Content.Projectiles.Enemies
@@ -61,9 +62,12 @@ namespace ArcaneOdyssey.Content.Projectiles.Enemies
 				}
 			}
 
-			if (Penetrations == 2)
+			if (Penetrations == 2 && Main.myPlayer == Projectile.owner)
 			{
 				Projectile.Kill();
+				if (Main.netMode != NetmodeID.SinglePlayer)
+					NetMessage.SendData(MessageID.KillProjectile, -1, -1, null, Projectile.identity, Projectile.owner);
+				return;
 			}
 
 			Projectile.rotation = Projectile.velocity.ToRotation();
